changes should actually not have any impact on Persistence interface ...
