+++ apg-patch-service-common/src/main/java/com/apgsga/microservice/patch/api/PatchPersistence.java
@@ -36,6 +36,4 @@ public interface PatchPersistence {
 	public void clean();
 
 	public void init() throws IOException;
-	
-	public List<Patch> findWithObjectName(String objectName);
 }
